+++ lib/bm/instrumentations/puma/collector.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'socket'
+require 'puma'
 require 'prometheus/client'
 require 'tcp_server_socket_backlog/tcp_server_socket_backlog'
 
@@ -21,6 +22,8 @@ module BM
 
           # TCP_INFO with backlog statistics is Linux only
           @is_socket_backlog = TCPServer.method_defined?(:socket_backlog)
+
+          metrics_collection.server_version(::Puma::Server::VERSION)
         end
 
         # Updates Puma metrics in the registry