+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Security:
 
 Features:
 
+- `BM::Instrumentations::Puma` exports the running Server version as `puma_server_version(version)` gauge
+
 Bugfixes:
 
 ## 0.1.0 (2021-05-13)